Reworked Ballot ArrayList & Minor Fixes
The Ballot ArrayList was reworked to contain references to candidates instead of their names. This allows the ballot to add itself to a Candidate's vote list if its choice was advanced, which is used for the wonSeat method to advance all ballots that were not used for their chosen candidate's victory. Also added checks for some array and arrayList accessor methods, made the Candidate wonSeat method return a boolean based on whether it won a seat or not, and added an equals method to candidate that checks based on the candidate's name.