Sort and Remove candidateList
The candidateList is now sorted every time it is modified.  This will make it so that the candidate with the most votes overall will win the seat before the votes are redistributed, and if there is a tie it will be resolved via total votes. Candidate is also now removed from the candidate list once they win.