Fixed TotalVotes bug
The bug causing totalvotes to not count correctly has been fixed. There was no line that reset the candidateAdded flag after one was added, so this caused only 1 candidate to ever be added per ballot. causing this flag to reset after moving on to the next candidate has solved the issue. Debug lines were commented out